Create persian-new-variant-policy.py
According to the new IDNA standard IDNA2008, IRNIC is trying to adopt new registration policy to adopt new regulations. This code gets a domain-name (label) as in INPUT and try to convert all the variants. The same methodology applies for other languages and scripts. This project is trying to adopt this methodology and combine it with XML formating to achieve general plugin for every language.